New linux5.15-macg5 package with G5-specific dotconfig.
Initial version of a new package copied from linux5.15, but with
a customized .config file for PowerMac G5 systems, along with a
patch to enable setting "-mcpu=970" optimization from that .config.

The .config will likely require some fine-tuning to add any devices
that I removed that are present on other G5 systems but not mine,
as well as any PCI cards or filesystems that I should add back.
Feedback welcome!